Add BF16 support for int8_vectorwise_quant / LLM.int8 activation quant
Templates int8VectorQuant on T and adds bf16 kernel instantiations plus a
cint8_vector_quant_bf16 C ABI entry point, mirroring the existing
gemm_4bit_inference_naive fp16/bf16/fp32 pattern. The blockwise absmax
reduction now accumulates in float rather than T: required for bf16 to
compile cleanly and slightly improves fp16 accuracy (rowStats was already
float, so downstream is unaffected). Removes the forced A.to(torch.float16)
casts in MatMul8bitLt so bf16 activations quantize natively.

Closes #1868.

+++ bitsandbytes/autograd/_functions.py
@@ -134,10 +134,10 @@ class MatMul8bitLt(torch.autograd.Function):
         # 1. Quantize A. Note that as a side-effect, outliers are suppressed in CA/CAt.
         if ctx.needs_input_grad[1]:
             # Slower path
-            CA, CAt, SCA, SCAt, outlier_cols = F.int8_double_quant(A.to(torch.float16), threshold=state.threshold)
+            CA, CAt, SCA, SCAt, outlier_cols = F.int8_double_quant(A, threshold=state.threshold)
         else:
             # Fast path
-            CA, SCA, outlier_cols = F.int8_vectorwise_quant(A.to(torch.float16), threshold=state.threshold)
+            CA, SCA, outlier_cols = F.int8_vectorwise_quant(A, threshold=state.threshold)
             CAt = SCAt = None
 
         has_grad = False
@@ -152,7 +152,7 @@ class MatMul8bitLt(torch.autograd.Function):
                 state.reset_grads()
 
                 # 2. Quantize B
-                state.CB, state.SCB, _ = F.int8_vectorwise_quant(B.to(torch.float16))
+                state.CB, state.SCB, _ = F.int8_vectorwise_quant(B)
 
         # Handle sparse decomposition
         if state.threshold > 0.0:
@@ -258,7 +258,7 @@ class MatMul8bitFp(torch.autograd.Function):
 
             if (state.is_training and not has_grad) or state.CB is None or state.SCB is None:
                 state.reset_grads()
-                state.CB, state.SCB, _ = F.int8_vectorwise_quant(B.to(torch.float16))
+                state.CB, state.SCB, _ = F.int8_vectorwise_quant(B)
                 B = state.CB
 
         CB = state.CB.data.to(A.dtype).mul_(state.SCB.unsqueeze(1).mul(1.0 / 127.0))
